Modularize the multithreaded part of plugin image
