+++ applications/plugins/image/imageConfig.cmake.in
@@ -6,7 +6,6 @@
 set(IMAGE_HAVE_SOFA_GL @IMAGE_HAVE_SOFA_GL@)
 set(IMAGE_HAVE_SOFA_GUI_QT @IMAGE_HAVE_SOFA_GUI_QT@)
 set(IMAGE_HAVE_SOFAPYTHON @IMAGE_HAVE_SOFAPYTHON@)
-set(IMAGE_HAVE_MULTITHREADING @IMAGE_HAVE_MULTITHREADING@)
 set(IMAGE_HAVE_ZLIB @IMAGE_HAVE_ZLIB@)
 set(IMAGE_HAVE_FREENECT @IMAGE_HAVE_FREENECT@)
 
@@ -23,9 +22,6 @@ if(IMAGE_HAVE_SOFAGUIQT)
 endif()
 if(IMAGE_HAVE_SOFAPYTHON)
     find_package(SofaPython QUIET REQUIRED)
-endif()
-if(IMAGE_HAVE_MULTITHREADING)
-    find_package(MultiThreading QUIET REQUIRED)
 endif()
 if(IMAGE_HAVE_ZLIB)
     find_package(ZLIB QUIET REQUIRED)
