+++ applications/plugins/image/image_multithread/image_multithreadConfig.cmake.in
@@ -1,0 +1,7 @@
+# CMake package configuration file for the image_gui plugin
+
+@PACKAGE_GUARD@
+@PACKAGE_INIT@
+
+find_package(image QUIET REQUIRED)
+find_package(MultiThreading QUIET REQUIRED)
